Use named column lines instead of grid areas
If I used grid areas, I'll have to make an area for the header, an area
for the content, etc. I could have done that (and that will probably be
more predictable, since it's more explicit), but I'm liking this one.
Otherwise the header and content will stack on the same grid cell.