Make navbar responsive
There. I totally butchered it, I think. I didn't bother modifying the
gulpfile to watch the script, since I don't think I'll be working with
that too much.

Anyway this is a "quick" implementation of a responsive navbar. I'd like
to add a slide animation, but I have no idea how to do it "in a graceful
manner."